Remove redundant API tests #1005
