+++ vulnerabilities/tests/conftest.py
@@ -26,7 +26,6 @@ def no_rmtree(monkeypatch):
 # Step 3: Migrate all the tests
 collect_ignore = [
     "test_apache_kafka.py",
-    "test_api.py",
     "test_models.py",
     "test_package_managers.py",
     "test_ruby.py",
